Add form(e) equivalents for Species and Teammates.
- Form(e)s are treated as equivalent for the following species:
  - Maushold\*
  - Tatsugiri
  - Sinistcha
  - Poltchageist

No other cosmetic forms are distinguished on RK9 to my knowledge.

\* Re. Maushold:

Maushold has two formes, Family of Three (Maushold) and Family of Four
(Maushold-Four). These formes have a slightly different Weight which is
technically competitively relevant but only affects calcs against the
following Pokémon:

  - Turtwig using Heavy Slam
  - Tepig using Heat Crash
  - Rolycoly using Heat Crash
  - Appletun using Heavy Slam

Given that none of these Pokémon are competitively relevant and they
are too light to make effective use of those attacks in general, that
Maushold formes are used interchangably in practice, and that Maushold
is used consistently, I've decided that the value in querying Maushold
makes treating Maushold formes as interchangable a reasonable decision.